replace uplot ratio graph with chartjs graph
