+++ src/lib/components/RatioGraph.svelte
@@ -1,36 +1,33 @@
 <script lang="ts">
     import { onMount, onDestroy } from "svelte";
-    import { makeUserRatioOptions } from "$lib/utils/graph.ts";
+    import { makeUserRatioConfiguration } from "$lib/utils/graph.ts";
     import { queueRender } from "$utils/renderQueue";
-    import uPlot from "uplot";
-    import "uplot/dist/uPlot.min.css";
 
-    let {
-        timestamps,
-        values,
-        name,
-    }: { timestamps: number[]; values: number[]; name: string } = $props();
+    import type { Chart as ChartType } from "chart.js";
 
-    let graphContainer: HTMLDivElement;
-    let graphPlot: uPlot | undefined;
+    let Chart: any;
+
+    let { timestamps, values, name, is24h }: { timestamps: number[]; values: number[]; name: string; is24h?: Boolean } = $props();
+
+    let chartCanvas: HTMLCanvasElement;
+    let graphChart: ChartType | undefined;
     let resizeFrame: number | undefined;
     let lastWidth = 0;
     let mounted = false;
     let queuedRenderCancel: (() => void) | undefined;
 
-    const getData = (): uPlot.AlignedData => [timestamps, values];
-
     function createPlot() {
-        if (!mounted || graphPlot || !graphContainer?.isConnected) {
+        if (!mounted || graphChart || !chartCanvas?.isConnected) {
             return;
         }
 
-        lastWidth = graphContainer.clientWidth;
-        graphPlot = new uPlot(
-            makeUserRatioOptions(lastWidth, name, name),
-            getData(),
-            graphContainer,
-        );
+        lastWidth = chartCanvas.clientWidth;
+
+        graphChart = new Chart(chartCanvas, makeUserRatioConfiguration(timestamps, values, name, is24h));
+    }
+
+    function handleDoubleclick() {
+        graphChart?.resetZoom();
     }
 
     function recreatePlot() {
@@ -38,8 +35,8 @@
             return;
         }
 
-        graphPlot?.destroy();
-        graphPlot = undefined;
+        graphChart?.destroy();
+        graphChart = undefined;
         createPlot();
     }
 
@@ -51,17 +48,17 @@
         resizeFrame = requestAnimationFrame(() => {
             resizeFrame = undefined;
 
-            if (!graphPlot) {
+            if (!graphChart) {
                 return;
             }
 
-            const width = graphContainer.clientWidth;
+            const width = chartCanvas.clientWidth;
             if (width === 0 || width === lastWidth) {
                 return;
             }
 
             lastWidth = width;
-            graphPlot.setSize({ width, height: 400 });
+            graphChart.resize(width, 450);
         });
     }
 
@@ -71,43 +68,71 @@
     }
 
     $effect(() => {
-        if (graphPlot) {
-            graphPlot.setData(getData());
+        if (graphChart) {
+            graphChart.data.labels = timestamps;
+            graphChart.data.datasets[0].data = values;
+            graphChart.update();
         }
     });
 
     onMount(() => {
-        mounted = true;
+        let isCancelled = false;
 
-        const mediaQuery = window.matchMedia("(prefers-color-scheme: dark)");
-        const onSchemeChange = () => queueRender(recreatePlot);
+        let mediaQuery: MediaQueryList | undefined;
+        let onSchemeChange: (() => void) | undefined;
+        let ro: ResizeObserver | undefined;
+        let io: IntersectionObserver | undefined;
 
-        mediaQuery.addEventListener("change", onSchemeChange);
+        // dynamically import the modules on the client. zoom plugin will not work with ssr for example
+        const initChart = async () => {
+            const { Chart: ChartModule, registerables } = await import("chart.js");
+            const annotationPlugin = (await import("chartjs-plugin-annotation")).default;
+            const zoomPlugin = (await import("chartjs-plugin-zoom")).default;
 
-        const ro = new ResizeObserver(queueResize);
+            if (isCancelled) return;
 
-        ro.observe(graphContainer);
+            Chart = ChartModule;
+            Chart.register(...registerables, annotationPlugin, zoomPlugin);
 
-        const io = new IntersectionObserver(
-            ([entry]) => {
-                if (!entry?.isIntersecting) {
-                    return;
-                }
+            mounted = true;
 
-                queueCreatePlot();
-                io.disconnect();
-            },
-            { rootMargin: "300px" },
-        );
+            mediaQuery = window.matchMedia("(prefers-color-scheme: dark)");
+            onSchemeChange = () => queueRender(recreatePlot);
 
-        io.observe(graphContainer);
+            mediaQuery.addEventListener("change", onSchemeChange);
+
+            ro = new ResizeObserver(queueResize);
+
+            ro.observe(chartCanvas);
+
+            io = new IntersectionObserver(
+                ([entry]) => {
+                    if (!entry?.isIntersecting) {
+                        return;
+                    }
+
+                    queueCreatePlot();
+                    io?.disconnect();
+                },
+                { rootMargin: "300px" },
+            );
+
+            io.observe(chartCanvas);
+        };
+
+        initChart();
 
         return () => {
+            isCancelled = true;
             mounted = false;
             queuedRenderCancel?.();
-            mediaQuery.removeEventListener("change", onSchemeChange);
-            io.disconnect();
-            ro.disconnect();
+
+            if (mediaQuery && onSchemeChange) {
+                mediaQuery.removeEventListener("change", onSchemeChange);
+            }
+
+            io?.disconnect();
+            ro?.disconnect();
 
             if (resizeFrame != null) {
                 cancelAnimationFrame(resizeFrame);
@@ -116,13 +141,13 @@
     });
 
     onDestroy(() => {
-        graphPlot?.destroy();
-        graphPlot = undefined;
+        graphChart?.destroy();
+        graphChart = undefined;
     });
 </script>
 
-<div style="height: 500px; max-width: 700px; width: 100%">
-    <div bind:this={graphContainer} style="width: 100%"></div>
+<div style="height: 450px; max-width: 700px; width: 100%">
+    <canvas bind:this={chartCanvas} ondblclick={handleDoubleclick} style="width: 100%; height: 100%;"></canvas>
 </div>
 
 <style>
